% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvskPortfolio.R
\encoding{UTF-8}
\name{mvskPortfolio}
\alias{mvskPortfolio}
\title{Risk-based higher-order moment efficient portfolios}
\usage{
mvskPortfolio(m1 = NULL, M2 = NULL, M3 = NULL, M4 = NULL,
  w0 = NULL, g = NULL, kappa = NULL, lb = NULL, ub = NULL,
  lin_eq = NULL, lin_eqC = NULL, nlin_eq = NULL, lin_ieq = NULL,
  lin_ieqC = NULL, nlin_ieq = NULL, riskcriterion = function(w)
  sum((w - 1/p)^2), options = list(), relative = TRUE)
}
\arguments{
\item{m1}{mean vector; not used if NULL}

\item{M2}{covariance matrix}

\item{M3}{coskewness matrix; not used if NULL}

\item{M4}{cokurtosis matrix; not used if NULL}

\item{w0}{weight of the benchmark portfolio / initial portfolio name, see details}

\item{g}{vector with preferences for the moments, see details}

\item{kappa}{vector of values indicating maximum deviation of the benchmark portfolio}

\item{lb}{lower bound for the weights, default rep(0, p)}

\item{ub}{upper bound for the weights, default rep(1, p)}

\item{lin_eq}{equality constraints: eq w = eqC (should be matrix!), see details}

\item{lin_eqC}{equality constraints: eq w = eqC, see details}

\item{nlin_eq}{function with non-linear equality constraint (returns objective value and jacobian)}

\item{lin_ieq}{inequality constraints: ieq w leq ieqC (should be matrix!)}

\item{lin_ieqC}{inequality constraints: ieq w leq ieqC}

\item{nlin_ieq}{function with non-linear inequality constraints (returns objective value and jacobian)}

\item{riskcriterion}{optimal value of kappa minimizes the risk criterion function}

\item{options}{optimization options}

\item{relative}{steps of kappa are procentual if TRUE, absolute if FALSE}
}
\description{
function determining risk based portfolios that are mean-variance-skewness-kurtosis efficient.
}
\details{
moments with NULL are ignored; default is maxDiv as starting point; unconstrained, just set kappa to 1

possible initial portfolios are "maxDiv" for the maximum diversified portfolio of REF.

the vector g should have positive values, the default value is g = (0, sigma_^2_w0, abs(phi_w0), psi_w0)

lin_eq and lin_eqC default to the full investment constraint
}
\examples{
# load data
data(indices)
p <- 5
x <- as.matrix(indices[, 1:p])

# estimate moments
M2 <- cov(x)
M3 <- PerformanceAnalytics:::M3.MM(x, as.mat = FALSE)
M4 <- PerformanceAnalytics:::M4.MM(x, as.mat = FALSE)

# benchmark portfolio
w0 <- "maxDiv"

# optimal VSK portfolio
res <- mvskPortfolio(M2 = M2, M3 = M3, M4 = M4, w0 = w0)


}
\references{
Choueifaty, Y., & Coignard, Y. (2008). Toward maximum diversification.
Journal of Portfolio Management, 35(1), 40.

Briec, W., Kerstens, K., & Jokung, O. (2007). Mean-variance-skewness portfolio
performance gauging: a general shortage function and dual approach.
Management science, 53(1), 135-149.
}
\author{
Dries Cornilly
}
\concept{risk-based portfolio}
