% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvskPortfolio.R
\encoding{UTF-8}
\name{mvskPortfolio}
\alias{mvskPortfolio}
\title{Risk-based higher-order moment efficient portfolios}
\usage{
mvskPortfolio(m1 = NULL, M2 = NULL, M3 = NULL, M4 = NULL,
  w0 = NULL, g = NULL, lb = NULL, ub = NULL, lin_eq = NULL,
  lin_eqC = NULL, nlin_eq = NULL, lin_ieq = NULL, lin_ieqC = NULL,
  nlin_ieq = NULL, href = NULL, kappa = NULL, relative = FALSE,
  param = NULL, options = list(), mompref = NULL)
}
\arguments{
\item{m1}{mean vector; not used if NULL}

\item{M2}{covariance matrix}

\item{M3}{coskewness matrix; not used if NULL}

\item{M4}{cokurtosis matrix; not used if NULL}

\item{w0}{weight of the benchmark portfolio / initial portfolio name, see details}

\item{g}{vector with preferences for the moments, see details}

\item{lb}{lower bound for the weights, default rep(0, p)}

\item{ub}{upper bound for the weights, default rep(1, p)}

\item{lin_eq}{equality constraints: eq w = eqC (should be matrix!), see details}

\item{lin_eqC}{equality constraints: eq w = eqC, see details}

\item{nlin_eq}{function with non-linear equality constraint (returns objective value and jacobian)}

\item{lin_ieq}{inequality constraints: ieq w leq ieqC (should be matrix!)}

\item{lin_ieqC}{inequality constraints: ieq w leq ieqC}

\item{nlin_ieq}{function with non-linear inequality constraints (returns objective value and jacobian)}

\item{href}{reference function to keep into account while tilting the portfolios (name or function)}

\item{kappa}{vector of values indicating maximum deviation of the reference objective}

\item{relative}{boolean indicating if kappa is absolute or relative}

\item{param}{list with extra arguments for the href function}

\item{options}{optimization options}

\item{mompref}{direction of preference for the moments, defaults to higher mean and skewnes,
lower variance and kurtosis}
}
\description{
function determining risk based portfolios that are mean-variance-skewness-kurtosis efficient.
}
\details{
moments with NULL are ignored; default is maxDiv as starting point; unconstrained, just set kappa to 1

possible initial portfolios are "maxDiv" for the maximum diversified portfolio of REF.

the vector g should have positive values, the default value is g = (0, sigma_^2_w0, abs(phi_w0), psi_w0)

lin_eq and lin_eqC default to the full investment constraint
}
\examples{
# load data
library(PerformanceAnalytics)
data(edhec)
x <- edhec[, 1:5]

# estimate moments
m1 <- colMeans(x)
M2 <- cov(x)
M3 <- PerformanceAnalytics:::M3.MM(x, as.mat = FALSE)
M4 <- PerformanceAnalytics:::M4.MM(x, as.mat = FALSE)

# MVSK tilting - starting from "DR", tilting with margin on "DR"
resMVSK <- mvskPortfolio(m1 = m1, M2 = M2, M3 = M3, M4 = M4, w0 = "DR",
                         g = "mvsk", ub = rep(0.3, 5), href = "DR",
                         kappa = c(0, 0.01, 0.025, 0.05))

# show weights
barplot(resMVSK$w, beside = TRUE)

# VSK tiltin - starting from "DR", tilting with maximum tracking error
# volatility
resVSK <- mvskPortfolio(M2 = M2, M3 = M3, M4 = M4, w0 = "DR",
                        g = "mvsk", ub = rep(0.3, 5), href = "TEvol",
                        kappa = c(0, 0.001, 0.025))

# show weights
barplot(resVSK$w, beside = TRUE)

}
\references{
Boudt, K., Cornilly, D., Van Holle, F., & Willems, J. (2019). Algorithmic portfolio
tilting to harvest higher moment gains. working paper

Briec, W., Kerstens, K., & Jokung, O. (2007). Mean-variance-skewness portfolio
performance gauging: a general shortage function and dual approach.
Management science, 53(1), 135-149.

Choueifaty, Y., & Coignard, Y. (2008). Toward maximum diversification.
Journal of Portfolio Management, 35(1), 40.
}
\author{
Dries Cornilly
}
\concept{risk-based portfolio}
